requirements.txt
tweepy==4.14.0
python-dotenv==0.19.2
requests==2.32.3
